Create NthPersonGetsNthSeat.cs
This problem is to find the probability that the nth person will be able to get the correct seat. This problem doesn't take into account conditional probability. It is just asking basic probability, of whether the last person with one seat left what would be the probability that it is the correct seat.