Create minMoves2.java
This problem is asking to find the minimum movement to get the entire array equal to one another. This solution find the median value by sorting the list, and then subtracting each ends by the other until you each the median value. This is because the median value will be the minimum distance from the two end values.