Create climbStairs.cpp
This problem is asking how many ways there is climb n stairs if you are allowed to take 1 or 2 steps. This is a Fibonacci sequence that can be solved through dynamic programming while maintaining space by keeping only two items at a time rather than the entire list of n numbers. This is because the dependency is only for the two previous terms.