Create smallerNumbersThanCurrent.java
This question is asking to return an array that contains the number of items that are smaller than the item in the array. The brute force solution would be O(N^2) because you would have to loop through the entire list multiple times keeping a count of items that are below it. However, to optimize on speed not on space this solution keeps track of a frequency counter and then finds the number of items smaller than itself. This solution is O(N) in space as well as O(N) in time complexity.