Create hasCycle.java
This question is asking given a linked list whether it contains a cycle or not. The linear solution for this problem is to run through the linked list and if you reached the size limit there must have been a cycle. This is the first solution I came up with.